libhwc2.1: Duplicate acquireFence when it is passed
In Android Q, acquireFence is not passed when the buffer isn't
updated. So this thing can make some problems about buffer synchronization.
So this patch made modify that all of the layers will duplicate acquireFence
and store in mPrevAcquireFence.

If the layer isn't updated, fence that is duplicated from mPrevAcquireFence
will be used.

This 'mPrevAcquireFence' will be closed when the buffer is updated.

Bug: 190864548
Test: monkey, suspend/resume,AoD in/out
Change-Id: I3efe14a9517c6f0cab0af66c55286aeacde6a1bd
Signed-off-by: Mingu Jeon <mingu85.jeon@samsung.com>