gralloc3: support formats where planes are contiguous in memory
The current gralloc implementation assumes that alignment operations
need to be performed per-plane in multiplanar formats as they are
in separate buffers.

When this isn't the case, and the chroma planes follow immediately
after the luma plane, this can result in miscalculated stride values
for some buffer dimensions resulting in visual corruption.

Bug: 150319746
Change-Id: Ic8ec2e1069cf9dc99b28584a586c72be6fb8fb0c