libhwc2.1: Add display capability to skip client color tranform
If getDisplayCapabilities is supported, then the global capability
HWC2_CAPABILITY_SKIP_CLIENT_COLOR_TRANSFORM is ignored. So, we
need to add HWC2_DISPLAY_CAPABILITY_SKIP_CLIENT_COLOR_TRANSFORM
to ExynosDisplay::getDisplayCapabilities() which is similar to
HWC2_CAPABILITY_SKIP_CLIENT_COLOR_TRANSFORM in
ExynosDevice::getCapabilities().

This is to ensure that the client color transform will not be
applied twice which is causing the display color looks reddish
in night light mode when it's composited by client.

Bug: 190483828
Test: enable night light, the display color should not abnormally
      flash reddish color
Change-Id: I37c46142d5d311ac120256eb72fbd0a63fb21bb2